app/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-openai
langchain-litellm
langchain-community
langchain-chroma
chromadb
fastembed
litellm
psycopg2-binary
sqlalchemy
python-dotenv
pydantic
asyncpg
